Fixed bug where hitting the back button after visiting a particular blog post would render JSON instead of the previous page.
The root cause of the problem was that we were using the same endpoint to serve the HTML and JSON for the post page. That is,
when we rendered the page, we would make a call to the endpoint for the HTML and then make a subsequent call to the same
endpoint for JSON data to fill the page. The problem with this is that Chrome caches the response from the last call to a
particular endpoint without regard to the request's `Content-Type` header. As such, hitting the back button would cause the
browser to serve the last JSON response.

The fix is to make it so we just embed the data for the initial page load into the HTML payload. This way we avoid this bug
*and* avoid making extraneous calls to the server to get additional data.